move Spreadsheets to directory
