+++ appUI/src/pages/Compounds/UpsertCompoundDialog.tsx
@@ -26,7 +26,7 @@ import { useNotifications } from '../../contexts/NotificationContext.tsx';
 import { DialogTitleWithX } from '../../components/Mui/DialogTitle.tsx';
 import { ErrorLabel, InfoLabel, WarningLabel } from '../../components/dataViews/TooltipIcons.tsx';
 import { RestoreButton } from '../../components/Mui/RestoreButton.tsx';
-import { ParamsSpreadsheet } from '../../components/ParamsSpreadsheet.tsx';
+import { ParamsSpreadsheet } from '../../components/Spreadsheet/ParamsSpreadsheet.tsx';
 import { DialogProps } from '../../adapters/types/DialogProps.ts';
 import { InputVaporFitDialog } from './InputVaporFitDialog.tsx';
