+++ appUI/src/pages/Fitting/SpecifyFitDialog.tsx
@@ -24,7 +24,7 @@ import { ErrorLabel, InfoLabel } from '../../components/dataViews/TooltipIcons.t
 import { DialogProps } from '../../adapters/types/DialogProps.ts';
 import { SystemIdentifier } from '../../adapters/api/types/common.ts';
 import { PersistedFit } from '../../adapters/api/types/fitTypes.ts';
-import { ParamsSpreadsheet } from '../../components/ParamsSpreadsheet.tsx';
+import { ParamsSpreadsheet } from '../../components/Spreadsheet/ParamsSpreadsheet.tsx';
 import { useFitResultsDialog } from '../../actions/Fit/useFitResultsDialog.tsx';
 
 type UpsertDatasetDialogProps = DialogProps &
